Checkpoint

+++ app/forms/transaction_form.py
@@ -8,7 +8,7 @@ class RequestTransactionForm(FlaskForm):
     sender_id = StringField("sender")
     receiver_id = StringField("receiver")
     
-    request_amount = IntegerField("Request Amount")
+    request_amount = DecimalField("Request Amount")
     pay_amount = DecimalField("Pay Amount")
     is_Pending = BooleanField("Pending")
 
@@ -17,6 +17,6 @@ class SendTransactionForm(FlaskForm):
     sender_id = StringField("sender")
     receiver_id = StringField("receiver")
     receiver_username = StringField("receiver", validators=[DataRequired()])
-    request_amount = IntegerField("Request Amount")
+    request_amount = DecimalField("Request Amount")
     pay_amount = DecimalField("Pay Amount")
     is_Pending = BooleanField("Pending")
